more apt module names
